+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 > Read detailed change logs in https://github.com/sakurawald/fuji/commits/dev/
 
-- [fuji] fix: should not display the `static` fields in `/fuji inspect configurations` command.
+- [fuji] fix: should not display the `static` fields in `/fuji inspect configurations` command. (To hides the `not interested` fields.)
+- [core] fix: should remember the `parent GUI` when using the `search function` in `paged GUI`. (This makes the GUI interaction more convenient.)